refactor: update onboarding view and create reusable widgets
