+++ lib/features/onboarding/screen/widgets/onboarding_button.dart
@@ -1,0 +1,32 @@
+import 'package:dam_project/utils/constants/app_colors.dart';
+import 'package:dam_project/utils/device/device_utils.dart';
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+class OnBoardingButton extends StatelessWidget {
+  const OnBoardingButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(8),
+        color: AppColors.primary,
+      ),
+      width: DeviceUtils.getScreenWidth(context) * .9,
+      height: 55,
+      child: TextButton(
+        onPressed: () async {
+          final prefs = await SharedPreferences.getInstance();
+          prefs.setBool("onboarding", true);
+
+          Navigator.pushReplacement(
+            context,
+            MaterialPageRoute(builder: (context) => Placeholder()),
+          );
+        },
+        child: const Text("Get started", style: TextStyle(color: Colors.white)),
+      ),
+    );
+  }
+}
